requirements.txt
fastapi>=0.95.0
websockets>=11.0.0
prompt-toolkit>=3.0.0
pyshorteners>=1.0.1
uvicorn
pydantic
asyncio>=3.4.3
rich
typer
click
termcolor
pyinquirer
tqdm
colorama
pyfiglet
gql
requests
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests-toolbelt
python-dotenv
slack-sdk>=3.19.0
mem0ai
